+++ src/main/java/step2/domain/lotto/LottoNumberAutoGenerationStrategy.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2.domain.lotto;
 
 import step2.vo.LottoNumberValue;
