+++ src/main/java/step2/domain/lotto/Lottos.java
@@ -1,6 +1,4 @@
-package step2.domain;
-
-import step2.domain.Lotto;
+package step2.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
