+++ src/test/java/step2/domain/lotto/WinningLottoTest.java
@@ -1,7 +1,8 @@
-package step2.domain;
+package step2.domain.lotto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.statistics.WinningStatistics;
 import step2.utils.WinningStatisticsTestUtil;
 
 import java.util.Arrays;
